% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_models.R
\name{assemble_models}
\alias{assemble_models}
\title{ASSEMBLING MODELS FUNCTION}
\usage{
assemble_models(
  modelsList,
  method = c("ACC", "Jaccard Similarity", "TSS", "AUC", "kappa"),
  threshold = 0.5,
  modelspath = "./"
)
}
\arguments{
\item{modelsList}{List of bootstrap niche models}

\item{method}{String indicating method to weight bootstrap performance. Default "ACC"}

\item{threshold}{Threshold to select models cut-off. Default is 0.5}

\item{modelspath}{File path to the folder where bootstrap models can be found}
}
\value{
Ensemble NINA model
}
\description{
Assemble bootstrap models
}
\details{
Returns an error if \code{filename} does not exist.
}
