% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_snm_bootstraps.R
\name{assemble_snm_bootstraps}
\alias{assemble_snm_bootstraps}
\title{ASSEMBLING BOOTSTRAP MODELS}
\usage{
assemble_snm_bootstraps(
  z,
  env.scores,
  sp.scores,
  w = FALSE,
  bootstrap.eval = NULL,
  eval = F,
  threshold = 0.5,
  cluster = F,
  method = "ACC"
)
}
\arguments{
\item{z}{List of niche models}

\item{env.scores}{Envirronmental PCA scores}

\item{sp.scores}{Occurences environmental scores}

\item{w}{Boolean indicating the use of weigthing coefficients}

\item{bootstrap.eval}{List of model evaluations}

\item{eval}{Boolean indicating if use the model evaluations}

\item{threshold}{Numeric thesthold to select models cut-off. Default 0.5}

\item{cluster}{Boolean if models have been clustered into regions}

\item{method}{String indicating which parameter to use to weight models performance. Default "ACC"}
}
\value{
Niche model
}
\description{
Assemble bootstrap niche models
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
