% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_snm_models.R
\name{assemble_snm_models}
\alias{assemble_snm_models}
\title{ASSEMBLING MODELS FUNCTION}
\usage{
assemble_snm_models(
  modelsList,
  type = c("EN", "BC", "EC"),
  method = c("ACC", "Jaccard Similarity", "TSS", "AUC", "kappa"),
  threshold = 0.5,
  modelspath = "./"
)
}
\arguments{
\item{modelsList}{List of bootstrap niche models}

\item{type}{String indicating the type of niche model}

\item{method}{String indicating method to weight bootstrap performance. Default "ACC"}

\item{threshold}{Threshold to select models cut-off. Default is 0.5}

\item{modelspath}{File path to the folder where bootstrap models can be found}
}
\value{
Data frame.
}
\description{
Assemble bootstrap models
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
