% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_model.R
\name{BC_model}
\alias{BC_model}
\title{ESTIMATE BC MODEL}
\usage{
BC_model(
  x,
  y,
  A.matrix = NULL,
  C.matrix = NULL,
  D = 1,
  method = c("composition", "densities"),
  eval = TRUE,
  relative.niche = T,
  K = NULL,
  sample.pseudoabsences = TRUE,
  R = 100,
  res = NULL,
  plot.eval = FALSE,
  rep = 100,
  th = NULL,
  ras = NULL,
  best.th = c("accuracy", "similarity"),
  combine.regions = F,
  cor = F,
  type = c("region", "global")
)
}
\arguments{
\item{x}{NINA EN model object for species group one}

\item{y}{NINA EN model object for species group two}

\item{A.matrix}{m by n matrix indicating the association coefficient (-1 to 1). m are species to be modeled as rows and n interactions as columns}

\item{C.matrix}{n by n matrix indicating the competition coefficient between interactions (0 to 1).}

\item{D}{Numeric value indicating independence from biotic associations. Value must be comprised between 0 and 1.}

\item{method}{Method; abundances or composition}

\item{eval}{Boolean whether to evaluate the model}

\item{relative.niche}{Logical}

\item{K}{= Carrying capacity of each environmental cell}

\item{sample.pseudoabsences}{Boolean to whether sample pseudo-absences}

\item{R}{Integer. size of the grid for the niche space estimate. R represents number of columns and rows}

\item{res}{spatial resolution}

\item{plot.eval}{Logical to whether plot the evaluation}

\item{rep}{number of randomzation tests}

\item{th}{threshold to perform cut off for model evaluation}

\item{ras}{raster to constrain pseudoabsences sampling in model evalluation}

\item{best.th}{method to select the best thresholt. Default is "similarity"}

\item{combine.regions}{Logical. Whether to combine regional niche models into a global one}

\item{cor}{Logical}

\item{type}{String indicating whether to perform at a region or a global level. Note that if models have not been estimated at a region level and it is selected it will produce an error}
}
\value{
NINA model
}
\description{
Transforms environmental niche in base to species interactions
}
\examples{
\dontrun{
EN1 <- EN_model(env_data, occ_data1, cluster = "env", n.clus = 5)
EN2 <- EN_model(env_data, occ_data2, cluster = "env", n.clus = 5)
BC <- BC_model(EN1, EN2, A.matrix = int_matrix)
}

}
