% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_model_.R
\name{BC_model_}
\alias{BC_model_}
\title{ESTIMATE BC MODEL internal function}
\usage{
BC_model_(z, y.list, id, D = 0, A.matrix = NULL, C.matrix = NULL)
}
\arguments{
\item{z}{environmental niche model}

\item{y.list}{list of environmental niche models of species interactors}

\item{id}{species name to be estimated. id must match the rownames in A.matrix}

\item{D}{Binary integer indicating species independence on interactions. 0 for fully dependent, 1 for species that can be present without interactors}

\item{A.matrix}{m by n matrix indicating the association coefficient (-1 to 1). m are species to be modelled as rows and n interactors as columns}

\item{C.matrix}{n by n matrix indicating the competition coefficient between interactors (0 to 1).}
}
\value{
Data frame.
}
\description{
Transforms environmental niche in base to species interactions
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
