% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_regions.R
\name{cluster_regions}
\alias{cluster_regions}
\title{CLUSTERING FUNCTION}
\usage{
cluster_regions(
  x,
  plot = FALSE,
  n.clus = NULL,
  nstart = 25,
  K.max = NULL,
  B = 100,
  res = NULL
)
}
\arguments{
\item{x}{data frame to be clustered}

\item{plot}{Boolean whether to plot the cluster as map}

\item{n.clus}{Number of clusters to perform. If null it will be computed}

\item{nstart}{burn up start value}

\item{K.max}{maximum number of clusters to estimate}

\item{B}{number of runs}

\item{res}{Resolution of the greographical grid}
}
\value{
Clustered data frame.
}
\description{
Cluster a data frame with two first columns indicating long and lat coordinates
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
