% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_regions.R
\name{combine_regions}
\alias{combine_regions}
\title{COMBIINE REGIONAL MODELS FUNCTION}
\usage{
combine_regions(z.list, env.scores, R = 100)
}
\arguments{
\item{z.list}{List of regional niche models}

\item{env.scores}{Environmental PCA scores}

\item{R}{niche grid}
}
\value{
Niche model
}
\description{
Combine regional niche models
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
