% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EN_model.R
\name{EN_model}
\alias{EN_model}
\title{EN MODELLING FUNCTION}
\usage{
EN_model(
  env,
  occ,
  res = NULL,
  path = "./",
  project.name = "NINA_EN",
  nstart = 25,
  k.max = NULL,
  B = 100,
  crs = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0",
  extrapolate.niche = FALSE,
  save.bootstraps = F,
  save.model = F,
  relative.niche = T,
  combine.clusters = FALSE,
  cluster = NULL,
  n.clus = NULL,
  R = 100,
  sample.pseudoabsences = TRUE,
  eval = FALSE,
  split.data = FALSE,
  split.percentage = 0.25,
  split.method = c("kmeans", "Euclidean"),
  bootstraps = 1,
  assemble.models = TRUE,
  assemble.method = c("ACC", "Jaccard Similarity", "TSS", "AUC", "kappa")
)
}
\arguments{
\item{env}{Environmental data frame}

\item{occ}{Occurrence dataset}

\item{res}{grid resolution of the spatial extent}

\item{path}{Directory path of partition models}

\item{project.name}{String indicating the project name. Default is "NINA_EN"}

\item{nstart}{Burn up start for clustering estimation}

\item{k.max}{Maximum number of cluseters}

\item{B}{Number of runs}

\item{crs}{CRS object or a character string describing a projection and datum in PROJ.4 format}

\item{extrapolate.niche}{Boolean wheter to allow niche extrapolation if needed}

\item{save.bootstraps}{Boolean whether to save partitions}

\item{save.model}{Boolean whether to save the model}

\item{relative.niche}{logical. Only in case of using clustering method. If TRUE, computes the relative species niche density over the overall species niche clusters.}

\item{combine.clusters}{Boolean whether to combine regions into global models}

\item{cluster}{data frame with clustered regions as levels}

\item{n.clus}{number of clusters to perform}

\item{R}{niche space grid}

\item{sample.pseudoabsences}{Boolean to indicate if pseudoabsences are to be sampled}

\item{eval}{Boolean whether to evaluate the model}

\item{split.data}{Boolean whether to split data in case of model bootstrapping}

\item{split.percentage}{Split percentage, Default is 0.25}

\item{split.method}{String indicating the method to split occurrence data. Default is kmeans}

\item{bootstraps}{Integer indicating number of partitions to perform. Default is 1}

\item{assemble.models}{Boolean whether to assemble all partitions into ensemble model}

\item{assemble.method}{String indicating the evaluation parameter to weight and compute the assembling. Default is "ACC"}
}
\value{
List of elements
}
\description{
Samples pseudo-absences from a environmental extent based on euclidean distances to the observations in a environmental PCA
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
EN<- EN_model(env_data, occ_data1, boot)
}
\dontrun{
EN<- EN_model(env_data, occ_data2, cluster = "env", n.clus = 5)
}

}
