% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EN_model.R
\name{EN_model_}
\alias{EN_model_}
\title{BIN MODELLING FUNCTION}
\usage{
EN_model_(
  env,
  occ,
  res = NULL,
  sample.pseudoabsences = TRUE,
  extrapolate.niche = FALSE,
  nstart = 25,
  k.max = NULL,
  B = 100,
  combine.clusters = FALSE,
  cluster = NULL,
  n.clus = NULL,
  R = 100,
  eval = FALSE,
  split.data = FALSE,
  split.percentage = 0.25,
  split.method = c("kmeans", "Euclidean")
)
}
\arguments{
\item{env}{Environmental data frame}

\item{occ}{Occurrence dataset}

\item{res}{grid resolution of the spatial extent}

\item{sample.pseudoabsences}{Boolean to indicate if pseudoabsences are to be sampled}

\item{extrapolate.niche}{Boolean wheter to allow niche extrapolation if needed}

\item{nstart}{Burn up start for clustering estimation}

\item{k.max}{Maximum number of cluseters}

\item{B}{Number of runs}

\item{combine.clusters}{Boolean whether to combine regions into global models}

\item{cluster}{data frame with clustered regions as levels}

\item{n.clus}{number of clusters to perform}

\item{R}{niche space grid}

\item{eval}{Boolean whether to evaluate the model}

\item{split.data}{Boolean wheter to split data in case of model boostrapping}

\item{split.percentage}{Split percentage, Default is 0.25}

\item{split.method}{String indicating the method to split occurrence data. Default is kmeans}
}
\value{
List of elements
}
\description{
Samples pseudo-absences from a environmental extent based on euclidean distances to the observations in a environmental PCA
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
