% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_data.R
\docType{data}
\name{env_data}
\alias{env_data}
\title{Environmental Sample Data for Analysis}
\format{
A \code{raster} object with 6 environmental layers, which are:
\describe{
\item{Bio1}{Annual Mean Temperature}
\item{Bio2}{Mean Diurnal Range}
\item{Bio3}{Isothermality}
\item{Bio4}{Temperature Seasonality}
\item{Bio5}{Max Temperature of Warmest Month}
\item{Bio6}{Min Temperature of Coldest Month}
}
}
\usage{
env_data
}
\description{
A \code{raster} sample of BioClim environmental variables for testing and running examples
}
\keyword{datasets}
