% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_betas.R
\name{estimate_betas}
\alias{estimate_betas}
\title{ESTIMATE BETAS}
\usage{
estimate_betas(y.list, C.matrix = NULL)
}
\arguments{
\item{y.list}{List of niche models}

\item{C.matrix}{Competition matrix}
}
\value{
list of lists
}
\description{
Estimate the beta values of interactors
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
