% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{EVALUATE FUNCTION}
\usage{
evaluate_model(
  Fit.,
  Obs.,
  th = NULL,
  rep = 1000,
  best.th = c("similarity", "accuracy"),
  main = "",
  plot = T
)
}
\arguments{
\item{Fit.}{Vector of predicted values}

\item{Obs.}{Vector of observed values}

\item{th}{threshold cut off}

\item{rep}{number of randomization tests}

\item{best.th}{method to select the best threshold. Default is "similarity"}

\item{main}{If plot = TRUE. Indicate plot title}

\item{plot}{Boolean whetehr to plot the result}
}
\value{
List
}
\description{
Evaluate independent model
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
