% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_matrix.R
\docType{data}
\name{int_matrix}
\alias{int_matrix}
\title{Sample if a simulated interaction matrix between two groups of species}
\format{
A \code{\link{matrix}} with Species Group One names as columns and Species Group Two names as rows.
}
\usage{
int_matrix
}
\description{
A \code{\link{matrix}} indicating species from two groups interact. 0 indicates no interactions and 1 indicates positive interaction. Column names indicate species name of Species Group One (See \code{\link{occ_data1}}). Row names indicate species names of Species Group Two (See \code{\link{occ_data2}})
}
\keyword{datasets}
