% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{LOAD NICHE MODEL}
\usage{
load_model(project.name, path = "~")
}
\arguments{
\item{project.name}{name of the folder to be created}

\item{path}{directory path}
}
\description{
Saves NINA niche model into the disk
}
\details{

}
