% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_comparison.R
\name{niche_comparison}
\alias{niche_comparison}
\title{NICHE COMPARISON}
\usage{
niche_comparison(
  x,
  y,
  centroid.w = F,
  rnd.test = F,
  Z_space = F,
  quantile = 0.75,
  np.type = c("unimodal", "multimodal"),
  np.metric = c("median", "mean", "max"),
  cor = F,
  rep = 100,
  rand = 1,
  alternative = c("greater", "lower")
)
}
\arguments{
\item{x}{NINA niche object}

\item{y}{NINA niche object}

\item{centroid.w}{logical whether to weight niche overlap by distance to niche centroid}

\item{rnd.test}{Logical to indicate if randomization tests are to be estimated}

\item{Z_space}{lLogical. Whether to compute USE parameters accounting for the environmental space (Z)}

\item{quantile}{If \code{centroid.w} is TRUE, quantle threshold to filter niche densities. Default is 0.5. See \code{\link[NINA]{niche_position}}}

\item{np.type}{If \code{centroid.w} is TRUE, type of centroid to estimate. Default is 'unimodal'. See \code{\link[NINA]{niche_position}}}

\item{np.metric}{If \code{centroid.w} is TRUE, statistic method to estimate the niche centroid. Default is 'median'. See \code{\link[NINA]{niche_position}}}

\item{cor}{Logical whether to use environmentally corrected densites}

\item{rep}{Number of random samples to carry out the statistic tests}

\item{rand}{Directions of the comparisons. 1 performs all tests with argument 'x' as base of all comparisons. 2 uses argument 'y'. Default is 1.}

\item{alternative}{String indicating the alternative hypothesis. Default is "greater".}
}
\value{
An S3 object class NINA containing different niche anallyses.
}
\description{
Computes different metrics to analyze differences between two  niches
}
