% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_overlap.R
\name{niche_overlap}
\alias{niche_overlap}
\title{Niche Overlap}
\usage{
niche_overlap(
  x,
  y,
  cor = F,
  centroid.w = F,
  type = "unimodal",
  method = "median",
  quantile = 0.95
)
}
\arguments{
\item{x}{NINA niche object}

\item{y}{NINA niche object}

\item{cor}{Logical whether to use environmentally corrected densites}

\item{centroid.w}{logical whether to weight niche overlap by distance to niche centroid}

\item{type}{If \code{centroid.w} is TRUE, type of centroid to estimate. Default is 'unimodal'. See \code{\link[NINA]{niche_position}}}

\item{method}{If \code{centroid.w} is TRUE, statistic method to estimate the niche centroid. Default is 'median'. See \code{\link[NINA]{niche_position}}}

\item{quantile}{If \code{centroid.w} is TRUE, quantle threshold to filter niche densities. Default is 0.5. See \code{\link[NINA]{niche_position}}}
}
\value{
Numeric value indicating the proportion of niche overlap between two niches
}
\description{
Estimates the proportion of niche overlap between two niche objects of class NINA. The function uses the Schoener's D metric with an optional implementation that weights niche differences in base to the proximity to the niche centroid of the first argument.
}
