% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_parameters.R
\name{niche_parameters}
\alias{niche_parameters}
\title{NICHE PARAMETERS}
\usage{
niche_parameters(EN, EC, type = c("region", "global"))
}
\arguments{
\item{EN}{NINA object. Environmental niche model}

\item{EC}{NINA object. Ecological niche model}

\item{type}{String indicating the scale of the analysis. Default is "region"}
}
\value{
Data frame.
}
\description{
Estimates niche parameters
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
