% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_position.R
\name{niche_position}
\alias{niche_position}
\title{NICHE POSITION}
\usage{
niche_position(
  z,
  env,
  type = c("unimodal", "multimodal"),
  method = c("max", "median", "mean"),
  quantile = 0.5,
  cor = FALSE
)
}
\arguments{
\item{z}{niche model}

\item{env}{environemtnal variables data frame}

\item{type}{type of estimation. Default is "unimodal"}

\item{method}{Method to estimate the niche centroid. Default is "max"}

\item{quantile}{Numeric quantile to filter the niche}

\item{cor}{Boolean to whether use cor or uncor estimation}
}
\value{
Data frame.
}
\description{
Vector
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
