% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_to_dis.R
\name{niche_to_dis}
\alias{niche_to_dis}
\title{Niche to distribution}
\usage{
niche_to_dis(env.scores, z, cor = FALSE, cluster = NULL, w = NULL)
}
\arguments{
\item{env.scores}{Envirronmental PCA scores}

\item{z}{niche model}

\item{cor}{Boolean}

\item{cluster}{data frame with clustered data}

\item{w}{Weighting coefficients}
}
\value{
data frame
}
\description{
Niche to distribution
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
