% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_data1.R
\docType{data}
\name{occ_data1}
\alias{occ_data1}
\title{Sample of Occurrence Data of Species Group One for Analysis}
\format{
A \code{\link{data.frame}} with three columns, which are:
\describe{
\item{x}{Spatial longitud coordinates}
\item{y}{Spatial latitude coordinates}
\item{species}{species names belonging to each long lat coordinate}
}
}
\usage{
occ_data1
}
\description{
A \code{\link{data.frame}} with spatial coordinates of Species Group One. Occurrences have been sampled using \code{\link[virtualspecies]{sampleOccurrences}} from generated niches using \code{\link[virtualspecies]{generateRandomSp}}.
}
\keyword{datasets}
