% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_niche_overlap.R
\name{pairwise_niche_overlap}
\alias{pairwise_niche_overlap}
\title{SPECIES MODEL PAIRWISE NICHE OVERLAP FUNCTION}
\usage{
pairwise_niche_overlap(
  x,
  transformed = F,
  mode = c("region", "global"),
  write.csv = T,
  filename = "pairwise.niche.overlap.csv",
  alternative = c("lower", "greater", "both"),
  rand.type = 1,
  cor = F,
  rep = 10
)
}
\arguments{
\item{x}{Niche model. NINA object}

\item{transformed}{Boolean whetehr use the environmental or ecoligcal niche}

\item{mode}{Scale of the estimation. Default is region}

\item{write.csv}{Boolean whether to write a csv file witht the results}

\item{filename}{Filename}

\item{alternative}{Type of statistic test. Default is "lower"}

\item{rand.type}{Type of randomization methos. Default is 1}

\item{cor}{Boolean whether to use cor or uncor estimation}

\item{rep}{number of randomization tests}
}
\value{
Data frame.
}
\description{
Estimate all pairwise niche overlap of modelled species
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
