% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.NINA.R
\name{plot.NINA}
\alias{plot.NINA}
\title{Plot}
\usage{
\method{plot}{NINA}(x, ...)
}
\arguments{
\item{x}{An object class NINA}

\item{...}{Additional arguments for S3 methods}
}
\value{
A \code{NINA} class object description
}
\description{
Plots the summary of the output models of an object class NINA
}
\examples{
\dontrun{
EN = EN_model(env_data, occ_data1, cluster = "env")
plot(EN)
}

}
