% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.NINA.R
\name{print.NINA}
\alias{print.NINA}
\title{Print}
\usage{
\method{print}{NINA}(x, ...)
}
\arguments{
\item{x}{An object class NINA}

\item{...}{Additional arguments for S3 methods}
}
\value{
A description of a \code{NINA} class object
}
\description{
Prints characteristics of an object class NINA
}
\examples{
\dontrun{
EN = EN_model(env_data, occ_data1, cluster = "env")
print(EN)
}
}
