% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_to_PA.R
\name{prob_to_PA}
\alias{prob_to_PA}
\title{Probabilities to Presence/Absence}
\usage{
prob_to_PA(x, th = NULL, output = c("data.frame", "raster"), ras = NULL)
}
\arguments{
\item{x}{raster or data frame object}

\item{th}{selected threshold to convert to P/A}

\item{output}{type of output objct wanted. data frame or raster object}

\item{ras}{User can provide a raster object as template to match resolution, grid size and projection}
}
\value{
Data frame or raster
}
\description{
Converts probability distributions into Presence/Absence
}
\details{
Returns an error if \code{x} does not exist.
}
