% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_projection.R
\name{raster_projection}
\alias{raster_projection}
\title{CREATE RASTER PROJECTIONS}
\usage{
raster_projection(
  df,
  spsNames = NULL,
  ras = NULL,
  crs = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"
)
}
\arguments{
\item{df}{predicted model data frame}

\item{spsNames}{species names to build maps from}

\item{ras}{raster template in case extent, resolution and CRS are to be set at once}

\item{crs}{CRS object or a character string describing a projection and datum in PROJ.4 format}
}
\value{
Raster stack
}
\description{
Create a stack raster map with species model predictions
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
