% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_niche.R
\name{read_niche}
\alias{read_niche}
\title{SAVE SPECIES NICHE}
\usage{
read_niche(filepath)
}
\arguments{
\item{filepath}{path to the file}
}
\description{
reads a species NINA niche into from disk
}
\examples{
\dontrun{
EN1 <- EN_model(env_data, occ_data1, cluster = "env", n.clus = 5)
write_niche(EN1$z.mod$A$spB, path = "./", file = "ENtest")
read_niche(path = "./", file = "ENtest.txt")
}

}
