% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_list.R
\name{reverse_list}
\alias{reverse_list}
\title{REVERSE LIST  FUNCTION}
\usage{
reverse_list(ll)
}
\arguments{
\item{ll}{list of lists}
}
\value{
List of lists
}
\description{
Reverts the subset of lists
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
