% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pseudoabsences.R
\name{sample_pseudoabsences}
\alias{sample_pseudoabsences}
\title{Sampling pseudo-absences}
\usage{
sample_pseudoabsences(
  Obs,
  predictors,
  spsNames = NULL,
  th = 0.95,
  ras = NULL,
  res = NULL,
  plot = F
)
}
\arguments{
\item{Obs}{a n x 3 data.frame class object indicating longitud, latitud and species in the three first columns}

\item{predictors}{a n x m data.frame class object indicating longitud, latitud and species in the two first columns. The rest of the columns must be the environmental variables from which the PCA will be computed. Alternatively a raster stack with environmental variables or a dudi.pca object. If dudi.pca, res argument must be provided, otherwhise it will be set to 1.}

\item{spsNames}{a string vector with selected species names to subset Obs. Default NULL will select all species present in the third column of paramteter Obs}

\item{th}{a numeric threshold to estimate pseudo-absence distances from observations environmental centroid. Default is 0.95}

\item{ras}{raster or stacked raster with the ecological contrains that will weight pseudo-absence probabilities. Default NULL}

\item{res}{resolution of the environmental grid. Default NULL means that will be computed automatically.}

\item{plot}{Boolean to indicate if computed distances is to be plotted}
}
\value{
List
}
\description{
Samples pseudo-absences from a environmental extent based on euclidean distances to the observations in a environmental PCA
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
EN = EN_model(env_data, occ_data1, cluster = "env", n.clus = 5)
PseudoA <- sample_pseudoabsences(EN)
}

}
