% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_model.R
\name{save_model}
\alias{save_model}
\title{SAVE NICHE MODEL}
\usage{
save_model(m, path = "~", project.name = "NINA1")
}
\arguments{
\item{m}{Niche model of class NINA}

\item{path}{directory path}

\item{project.name}{name of the folder to be created}
}
\description{
Saves NINA niche model into the disk
}
\details{

}
\examples{
\dontrun{
EN1 <- EN_model(env_data, occ_data1, cluster = "env", n.clus = 5)
save_model(EN1, path = "~/Desktop", project.name = "NINA1")
}

}
