% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{SPLIT OCCURRENCES FOR CROSS-VALIDATION FUNCTION}
\usage{
split_data(
  Obs,
  spsNames = NULL,
  method = c("Euclidean", "kmeans"),
  split.percentage = 0.25
)
}
\arguments{
\item{Obs}{a nx3 data.frame class object indicating longitud, latitud and species in the three first columns}

\item{spsNames}{Species names to split occurrences from}

\item{method}{Method used to split occurrences. Default is "Euclidean"}

\item{split.percentage}{split percentage between train adn test data. Default is 0.25}
}
\value{
List
}
\description{
Samples pseudo-absences from a environmental extent based on euclidean distances to the observations in a environmental PCA
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}


}
