% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.NINA.R
\name{summary.NINA}
\alias{summary.NINA}
\title{Summary}
\usage{
\method{summary}{NINA}(object, ...)
}
\arguments{
\item{object}{An object class NINA}

\item{...}{Additional arguments for S3 methods}
}
\value{
A list summarizing the features of a \code{NINA} class object
}
\description{
Summarizes the features of an object class NINA and prints a description
}
\examples{
\dontrun{
EN = EN_model(env_data, occ_data1, cluster = "env", n.clus = 5)
EN.summary = summary(EN)
}
\dontrun{
EN = EN_model(env_data, occ_data2)
summary(EN)
}
}
