% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_niche.R
\name{write_niche}
\alias{write_niche}
\title{SAVE SPECIES NICHE}
\usage{
write_niche(x, path = "~", file = "niche")
}
\arguments{
\item{x}{Niche model of class NINA}

\item{path}{directory path}

\item{file}{filename}
}
\description{
Writes a species NINA niche into the disk
}
\details{

}
\examples{
\dontrun{
EN1 <- EN_model(env_data, occ_data1, cluster = "env", n.clus = 5)
write_niche(EN1$z.mod$A$spB, path = "./", file = "ENtest")
}


}
