% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_model.R
\name{BC_model}
\alias{BC_model}
\title{ESTIMATE BC MODEL}
\usage{
BC_model(
  x,
  y,
  A.matrix = NULL,
  C.matrix = NULL,
  D = 1,
  method = c("composition", "densities"),
  relative.niche = T,
  K = NULL,
  cor = F,
  type = c("region", "global")
)
}
\arguments{
\item{x}{NINA EN model object for species group one}

\item{y}{NINA EN model object for species group two}

\item{A.matrix}{m by n matrix indicating the association coefficient (-1 to 1). m are species to be modeled as rows and n interactions as columns}

\item{C.matrix}{n by n matrix indicating the competition coefficient between interactions (0 to 1).}

\item{D}{Numeric value indicating independence from biotic associations. Value must be comprised between 0 and 1.}

\item{method}{Method; abundances or composition}

\item{relative.niche}{Logical}

\item{K}{= Carrying capacity of each environmental cell}

\item{cor}{Logical}

\item{type}{String indicating whether to perform at a region or a global level. Note that if models have not been estimated at a region level and it is selected it will produce an error}
}
\value{
Data frame.
}
\description{
Transforms environmental niche in base to species interactions
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
