% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_model.R
\name{EC_model}
\alias{EC_model}
\title{ESTIMATE EC MODEL}
\usage{
EC_model(BC, EN, D = 0, type = c("region", "global"))
}
\arguments{
\item{BC}{NINA EN or BC model}

\item{EN}{Optional. NINA EN Model}

\item{D}{Numeric value for independence of interactions}

\item{type}{String indicating whether to perform at a region or a global level. Note that if models have not been estimated at a region level and it is selected it will produce an error}
}
\value{
Data frame.
}
\description{
Transform environmental niche space into ecological niche space
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
