% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_model.R
\name{EC_model}
\alias{EC_model}
\title{ESTIMATE EC MODEL}
\usage{
EC_model(
  x,
  y,
  D = 1,
  A.matrix = NULL,
  cor = F,
  K = NULL,
  method = c("composition", "densities"),
  relative.niche = T,
  C.matrix = NULL,
  type = c("region", "global")
)
}
\arguments{
\item{x}{NINA EN or BC model}

\item{y}{Optional. NINA EN Model}

\item{D}{Numeric value for independence of interactions}

\item{A.matrix}{m by n matrix indicating the association coefficient (-1 to 1). m are species to be modeled as rows and n interactions as columns}

\item{cor}{Logical}

\item{K}{= Carrying capacity of each environmental cell}

\item{method}{Method; abundances or composition}

\item{relative.niche}{Logical}

\item{C.matrix}{n by n matrix indicating the competition coefficient between interactions (0 to 1).}

\item{type}{String indicating whether to perform at a region or a global level. Note that if models have not been estimated at a region level and it is selected it will produce an error}
}
\value{
Data frame.
}
\description{
Transform environmental niche space into ecological niche space
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
  g1_EN = EN_model(env_data, occ_data1,  cluster = "env", n.clus = 5, relative.niche = F, cor = F)
  g2_EN = EN_model(env_data, occ_data2,  cluster = g1_EN$clus, relative.niche = F, cor = F)
  g2_BC <- BC_model(g2_EN, g1_EN, A.matrix = int_matrix, C.matrix = NULL, type = "region")
  g2_EC <- EC_model(g2_BC, type = "region")
}

}
