% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_model.R
\name{EC_model_}
\alias{EC_model_}
\title{ESTIMATE EC MODEL}
\usage{
EC_model_(en, W, R)
}
\arguments{
\item{en}{Niche model ito transform niche space}

\item{W}{Weighting coefficients}

\item{R}{Niche space grid}
}
\value{
Data frame.
}
\description{
Transform environmental niche space into ecological niche space
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
