% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_niche.R
\name{estimate_niche}
\alias{estimate_niche}
\title{Spatial Abundance Density Index}
\usage{
estimate_niche(
  glob,
  glob1,
  sp,
  R,
  h = "href",
  mask = NULL,
  th.o = NULL,
  th.s = NULL,
  method = c("epa", "bivnorm")
)
}
\arguments{
\item{glob}{2-column coordinates matrix of the environmental space}

\item{glob1}{2-column coordinates matrix of the subset environmental space}

\item{sp}{2-column coordinates matrix of the species occurrences position on the environmental space}

\item{R}{extent of the raster to be created}

\item{h}{smoothing parameter for the kernel estimation. Default is 'href'. Alternatively can be set to 'LSCV' or any given numeric value}

\item{mask}{raster mask to resample the created kernel density grid raster}

\item{th.o}{numeric threshold to filter density values of occurrences}

\item{th.s}{numeric threshold to filter density values of environment}

\item{method}{"epanechnikov" or "bivnorm"}
}
\value{
list object class niche
}
\description{
Estimates the occurrence density index of a set of occurrences over a 2-D grid space
}
\examples{
\dontrun{
env <- cbind(runif(1000, -1, 5), runif(1000, -2, 10))
sp <- cbind(runif(10, -1, 5), runif(10, -2, 10))
niche<- estimate_niche(sp, ext = c(-1,5,-2,10))
ecospat::ecospat.plot.niiche(niche)
}

}
