% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_w.R
\name{estimate_w}
\alias{estimate_w}
\title{ESTIMATE ENVIRONMENTAL ACCESSIBILITY}
\usage{
estimate_w(
  y.list,
  id,
  A.matrix = NULL,
  cor = F,
  K = NULL,
  method = c("composition", "densities"),
  C.matrix = NULL
)
}
\arguments{
\item{y.list}{list of niche models}

\item{id}{species id}

\item{A.matrix}{Association matrix}

\item{cor}{Logical}

\item{K}{= Carrying capacity of each environmental cell}

\item{method}{Method; abundances or composition}

\item{C.matrix}{Competition matrix}
}
\value{
Data frame.
}
\description{
Estimates the omega
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
