% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_evaluation.R
\name{models_evaluation}
\alias{models_evaluation}
\title{MODEL EVALUATION FUNCTION}
\usage{
models_evaluation(
  Pred,
  Obs,
  predictors,
  spsNames = NULL,
  th = NULL,
  ras = NULL,
  int.matrix = NULL,
  sample.pseudoabsences = TRUE,
  res = NULL,
  plot = TRUE,
  rep = 100,
  best.th = c("accuracy", "similarity")
)
}
\arguments{
\item{Pred}{Predicted niche models}

\item{Obs}{Occurrence dataset}

\item{predictors}{Environmental predictors}

\item{spsNames}{species names to evaluate}

\item{th}{threshol to perform cut off}

\item{ras}{raster to constrain pseudoabsences sampling}

\item{int.matrix}{interaction matrix between species and interactors}

\item{sample.pseudoabsences}{Boolean to whether sample pseudo-absences}

\item{res}{spatial resolution}

\item{plot}{Boolean to whether plot the evaluation}

\item{rep}{number of randomzation tests}

\item{best.th}{method to select the best thresholt. Default is "similarity"}
}
\value{
List
}
\description{
full models evaluation
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
