% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_parameters.R
\name{niche_parameters}
\alias{niche_parameters}
\title{NICHE PARAMETERS}
\usage{
niche_parameters(
  EN,
  EC,
  type = c("region", "global"),
  quantile = 0.75,
  np.metric = c("median", "mean", "max")
)
}
\arguments{
\item{EN}{NINA object. Environmental niche model}

\item{EC}{NINA object. Ecological niche model}

\item{type}{String indicating the scale of the analysis. Default is "region"}

\item{quantile}{Numeric value defining quantile threshold to filter niche suitability}

\item{np.metric}{Statistic metric to compute the niche position. Defaullt is median}
}
\value{
Data frame.
}
\description{
Estimates niche parameters
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun{
accident_2015 <- fars_read("Project/data/accident_2015.csv.bz2")
}

}
