% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.NINA.R
\name{summary.NINA}
\alias{summary.NINA}
\title{Summary}
\usage{
\method{summary}{NINA}(x, ...)
}
\arguments{
\item{x}{An object class NINA}

\item{...}{Additional arguments for S3 methods}
}
\value{
A \code{NINA} class object description
}
\description{
Suummarizes the output of an object class NINA
}
\examples{
\dontrun{
EN = EN_model(env_data, occ_data1, cluster = "env")
summary(EN)
}
}
